refactoring an rv-and stage out of test area
