+++ lib/src/ready_valid.dart
@@ -147,3 +147,51 @@ class ReadyValidConnector extends Module {
     return downstream;
   }
 }
+
+/// Module implementing a ready-AND-valid protocol
+class ReadyAndValidStage extends Module {
+  /// External enable for data transfer
+  Logic get transfer => output('transfer');
+
+  /// True if bubblesquashing is performed (e.g., can accept a first data token
+  /// before a ready)
+  final bool bubbleSquashing;
+
+  /// Create a module implementing ready-AND-valid
+  ReadyAndValidStage(Logic clk, Logic reset, ReadyValidInterface upstream,
+      ReadyValidInterface downstream,
+      {this.bubbleSquashing = true, super.name = 'readyAndValid'}) {
+    clk = addInput('clk', clk);
+    reset = addInput('reset', reset);
+
+    upstream = ReadyAndValidInterface(
+      modify: (original) => 'up_$original',
+    )..pairConnectIO(this, upstream, PairRole.consumer);
+    downstream = ReadyAndValidInterface(
+      modify: (original) => 'dn_$original',
+    )..pairConnectIO(this, downstream, PairRole.provider);
+
+    addOutput('transfer');
+
+    transfer <= upstream.valid & upstream.ready;
+    if (bubbleSquashing) {
+      final f = Fifo(clk, reset,
+          depth: 1,
+          writeEnable: transfer,
+          readEnable: downstream.valid & downstream.ready,
+          writeData: upstream.data);
+      downstream.valid <= ~f.empty;
+      upstream.ready <= ~f.full | downstream.ready;
+      downstream.data <= f.readData;
+    } else {
+      // (a, b) <= flop((ap, bp));  Records in dart  (a,b)[1] <=
+      downstream.data <=
+          flop(clk, upstream.data,
+              en: transfer, reset: reset, resetValue: LogicValue.x);
+      downstream.valid <=
+          flop(clk, upstream.valid,
+              en: transfer, reset: reset, resetValue: LogicValue.x);
+      upstream.ready <= downstream.ready;
+    }
+  }
+}
